+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release Notes
 
-## 0.17.2 (2018-07-??)
+## 0.17.3 (2018-07-19)
+
+* Fixed "PIP: Command "python setup.py egg_info" failed" (issue [#179](https://github.com/platformio/platformio-vscode-ide/issues/179))
+* Fixed "TypeError: Cannot read property 'title' of undefined" (issue [#170](https://github.com/platformio/platformio-vscode-ide/issues/170))
+
+## 0.17.2 (2018-07-11)
 
 - Do not force PIO IDE Terminal to `cmd.exe` on Windows (issue [#76](https://github.com/platformio/platformio-vscode-ide/issues/76))
 - [PIO Unified Debugger](http://docs.platformio.org/page/plus/debugging.html): Fixed infinite loading of Peripheral Registers